+++ src/constants/modalTypes.ts
@@ -1,9 +1,4 @@
-export const CUSTOM_BOOK_STATUS = 'customBookStatus';
-export const FILTERING = 'filtering';
-export const LANGUAGE_SETTINGS = 'languageSettings';
-export const CUSTOM_BOOK_CATEGORY = 'customBookCategory';
 export const DATE_UPDATER = 'dateUpdater';
-export const EDIT_GOAL = 'editModal';
 
 export const COMMON_SLIDE_MENU_HEIGHT = 336;
 export const BIG_SLIDE_MENU_HEIGHT = 486;