+++ src/screens/AddCustomBook/CategoryChooser/styles.ts
@@ -2,6 +2,12 @@ import { StyleSheet } from 'react-native';
 import colors from '~styles/colors';
 
 export default StyleSheet.create({
+  container: {
+    backgroundColor: colors.primary_dark,
+    display: 'flex',
+    flex: 1,
+    height: '100%',
+  },
   wrapper: {
     height: 300,
     width: '100%',